show install cli notification when triggered from welcome screen
