+++ crates/install_cli/Cargo.toml
@@ -16,7 +16,10 @@ test-support = []
 
 [dependencies]
 anyhow.workspace = true
+client.workspace = true
 gpui.workspace = true
+release_channel.workspace = true
 smol.workspace = true
 util.workspace = true
 workspace-hack.workspace = true
+workspace.workspace = true
